CreateAndEditChildForm.cs was updated
Fixed bug which causes null pointer exception. The problem was that in "Create" mod we don't have InitImgPath.  So InitImgPath.Substring throwed exception.

+++ PicturesSoft/PicturesSoft/CreateAndEditChildForm.cs
@@ -126,9 +126,6 @@ namespace PicturesSoft
 
                 if(!SourceFullFileName.Equals(destFolderName))
                 {
-                    string oldFileName =
-                        InitImgPath.Substring(InitImgPath.LastIndexOf("\\") + 1);
-
                     if (!imgExtension.Equals(".png"))
                     {
                         Image img;
@@ -156,9 +153,15 @@ namespace PicturesSoft
                             File.Copy(SourceFullFileName, destFolderName);
                     }
 
-                    if(!SourceFullFileName.Contains(PredeterminedDestImgFolderPath)
-                        && !oldFileName.Equals(fileName))
-                        File.Delete(InitImgPath);
+                    if (WorkMode.WorkType.Equals(WorkModeType.Edit))
+                    {
+                        string oldFileName =
+                        InitImgPath.Substring(InitImgPath.LastIndexOf("\\") + 1);
+
+                        if (!SourceFullFileName.Contains(PredeterminedDestImgFolderPath)
+                            && !oldFileName.Equals(fileName))
+                            File.Delete(InitImgPath);
+                    }        
                 }     
             }
             else //oldMode
